Bug 404745 - Document and cleanup TRS RIO
Clean up the TRS utilities and provide clearer javadoc

The cleanup in the TRS utilities will allow an adopter to more easily
understand and reuse the TRS utilities as a starting point.

Introduced IResourceUtil to return a list of known resources to use to 
prime Base objects.

Change the launch configuration to use a workspace reference with the
project name so that the launch config can be run from the menu bar
without the project being selected.


Change-Id: Ie0c73439181a72952a9a254a16d16db9bcf8fc54
Signed-off-by: emah <ernest@ca.ibm.com>