Fixing oslc4j-cm-test and the launch configuration
oslc4j-cm-test didn't get the change from changemanagement to cm.
The launch configuration for oslc4j-cm-sample was setting the wrong registry URI, this should be taken from the pom.xml instead.

Note: a possible better solution is to configure the host, port, etc. in the web.xml so that they are easier to change from a distributed WAR file.